renamed root folder for the package
