+++ src/store/index.ts
@@ -5,7 +5,7 @@ import { cooperationStore } from '@/cooperation/store';
 import { localStorageStore } from '@/user/store/localstorage';
 import { housesStore } from '@/houses/store';
 import { userStore } from '@/user/store/user';
-import { pollsStore } from '@/polls/store';
+import { pollsStore } from '@/poll/store';
 import { invitationsStore } from '@/store/invitations';
 import { apartmentsStore } from '@/apartment/store/apartments';
 import { ownershipsStore } from '@/apartment/store/ownerships';
